agrego titulo a la pagina de subcategoria, arreglo error al eliminar un producto para no permitir que se borre la imagen default.

+++ site/controllers/productsController.js
@@ -118,6 +118,7 @@ module.exports = {
             productos: dbProduct.filter(producto => {
                 return producto.subcategory == subcategory;
             }),
+            subcategoryTitle: subcategory,
             subcategory: [subcategory]
         })
 
@@ -185,7 +186,9 @@ module.exports = {
             producto.discount = req.body.discount,
             producto.description = req.body.description;
                 if (typeof req.files[0] != 'undefined'){
-                fs.unlinkSync('public/images/productos/' + producto.image);
+                    if(producto.image[0] != 'default-image.png'){
+                        fs.unlinkSync('public/images/productos/' + producto.image);
+                    }
                 producto.image = [req.files[0].filename]
                 }
             }
@@ -236,7 +239,9 @@ module.exports = {
         /* ELIMINO IMAGEN */
         dbProduct.forEach(producto => {
             if (producto.id == idProduct) {
-                fs.unlinkSync('public/images/productos/' + producto.image)
+                if(producto.image[0] != 'default-image.png'){
+                    fs.unlinkSync('public/images/productos/' + producto.image)
+                }
             }
         })
